[RAILS-04] Setup the database migration for the User model with Devise
Decided to drop the existing table and recreate it as it is easier than modify the table to look like we want.
I've enabled most of the Devise modules as they seem like good features to have for a social media.
I also cleaned the model as it seemed to be copy-pasta from another project.